Extract FavouriteMusic into favourite_music.rs (examples/dialogue_bot)
